Create python package definitions to support non-editable installs (#72)
* Hides pybind11 module implementation details

Refactors the Python module to re-export public symbols from a private compiled backend.
This prevents shadowing of the package directory and simplifies the user interface.
Additionally, the wheel now explicitly declares the package.

* Adds fallback for extension module import

Adds a fallback mechanism to locate the extension module.

This allows the extension to be found in cases where it is installed
at the top level, instead of inside the package directory. This
accommodates alternative installation methods.

* Renames the pybind module

Renames the pybind module from `kiss_matcher` to `_kiss_matcher`.

This change ensures consistency with naming conventions
for internal modules in Python packages.